requirements.txt
Django~=3.2.22
Pillow==9.5.0
django-admin-sortable2==1.0.4
